+++ src/Filter/Runner.fs
@@ -6,10 +6,12 @@ module internal FilterRunner =
 
     let runFilter: RunPattern<FilterApplication<'InputEvent, 'OutputEvent, 'Dependencies, 'FilterValue>, 'InputEvent, 'OutputEvent, 'Dependencies> =
         fun run (FilterApplication application) ->
-            let beforeRun filterApplication: BeforeRun<'InputEvent, 'OutputEvent, 'Dependencies> =
-                fun app ->
-                    (patternLogger FilterBuilder.pattern app.LoggerFactory)
-                        .LogDebug("Configuration: {configuration}", filterApplication.FilterConfiguration)
+            let beforeStart filterApplication = BeforeStart (fun app ->
+                (patternLogger FilterBuilder.pattern app.LoggerFactory)
+                    .LogDebug("Configuration: {configuration}", filterApplication.FilterConfiguration)
+            )
+
+            let beforeRun _ = BeforeRun.empty
 
             application
-            |> PatternRunner.runPattern FilterBuilder.pattern FilterApplication.application beforeRun run
+            |> PatternRunner.runPattern FilterBuilder.pattern FilterApplication.application beforeStart beforeRun run